perf(hadamard): vectorize FWHT butterfly with ARM64 asm unrolled scalar kernel
Moves fwht() out of hadamard.go into platform files:
- hadamard_generic.go (//go:build !arm64): original scalar triple-loop
- hadamard_arm64.go + hadamard_arm64.s: ARM64 asm kernel

The ARM64 kernel uses a 4-pair unrolled inner loop (F0–F15) backed by a
2-pair tier for the h=2 remainder. All pairs within each tier are fully
independent FP dependency chains, letting the OoO engine on Apple M-series
(6+ FP pipelines) execute them concurrently. The h=1 stage uses pointer
walking to eliminate per-pair address computation overhead.

Result on Apple M5 Pro (n=10 interleaved runs, benchstat):
  Forward:  −20–22% ns/op (p=0.000)
  Inverse:  −19–24% ns/op (p=0.000)
  Geomean:  −21.4% ns/op

Note: Go ARM64 assembler does not support FADDD/FSUBD with 3-operand
vector arrangements (C_ARNG×C_ARNG→C_ARNG illegal combination). The
4-pair scalar unroll achieves equivalent ILP on Apple M CPUs.
